+++ packages/@ngtools/webpack/src/compiler_host.ts
@@ -187,8 +187,10 @@ export class WebpackCompilerHost implements ts.CompilerHost {
   }
 
   invalidate(fileName: string): void {
-    this._files[fileName] = null;
-    this._changedFiles[fileName] = true;
+    if (fileName in this._files) {
+      this._files[fileName] = null;
+      this._changedFiles[fileName] = true;
+    }
   }
 
   fileExists(fileName: string): boolean {