+++ packages/@ngtools/webpack/src/loader.ts
@@ -2,8 +2,10 @@ import * as path from 'path';
 import * as ts from 'typescript';
 import {AotPlugin} from './plugin';
 import {TypeScriptFileRefactor} from './refactor';
+import {LoaderContext, ModuleReason} from './webpack';
 
 const loaderUtils = require('loader-utils');
+const NormalModule = require('webpack/lib/NormalModule');
 
 
 function _getContentOfKeyLiteral(source: ts.SourceFile, node: ts.Node): string {
@@ -16,11 +18,137 @@ function _getContentOfKeyLiteral(source: ts.SourceFile, node: ts.Node): string {
   }
 }
 
+
+function _angularImportsFromNode(node: ts.ImportDeclaration, sourceFile: ts.SourceFile): string[] {
+  const ms = node.moduleSpecifier;
+  let modulePath: string | null = null;
+  switch (ms.kind) {
+    case ts.SyntaxKind.StringLiteral:
+      modulePath = (ms as ts.StringLiteral).text;
+      break;
+    default:
+      return [];
+  }
+
+  if (!modulePath.startsWith('@angular/')) {
+    return [];
+  }
+
+  if (node.importClause) {
+    if (node.importClause.name) {
+      // This is of the form `import Name from 'path'`. Ignore.
+      return [];
+    } else if (node.importClause.namedBindings) {
+      const nb = node.importClause.namedBindings;
+      if (nb.kind == ts.SyntaxKind.NamespaceImport) {
+        // This is of the form `import * as name from 'path'`. Return `name.`.
+        return [(nb as ts.NamespaceImport).name.text + '.'];
+      } else {
+        // This is of the form `import {a,b,c} from 'path'`
+        const namedImports = nb as ts.NamedImports;
+
+        return namedImports.elements
+          .map((is: ts.ImportSpecifier) => is.propertyName ? is.propertyName.text : is.name.text);
+      }
+    }
+  } else {
+    // This is of the form `import 'path';`. Nothing to do.
+    return [];
+  }
+}
+
+
+function _ctorParameterFromTypeReference(paramNode: ts.ParameterDeclaration,
+                                         angularImports: string[],
+                                         refactor: TypeScriptFileRefactor) {
+  if (paramNode.type.kind == ts.SyntaxKind.TypeReference) {
+    const type = paramNode.type as ts.TypeReferenceNode;
+    const decorators = refactor.findAstNodes(paramNode, ts.SyntaxKind.Decorator) as ts.Decorator[];
+    const decoratorStr = decorators
+      .map(decorator => {
+        const fnName =
+          (refactor.findFirstAstNode(decorator, ts.SyntaxKind.CallExpression) as ts.CallExpression)
+          .expression.getText(refactor.sourceFile);
+
+        if (angularImports.indexOf(fnName) === -1) {
+          return null;
+        } else {
+          return fnName;
+        }
+      })
+      .filter(x => !!x)
+      .map(name => `{ type: ${name} }`)
+      .join(', ');
+
+    if (type.typeName.kind == ts.SyntaxKind.Identifier) {
+      const typeName = type.typeName as ts.Identifier;
+      if (decorators.length > 0) {
+        return `{ type: ${typeName.text}, decorators: [${decoratorStr}] }`;
+      }
+      return `{ type: ${typeName.text} }`;
+    }
+  }
+
+  return 'null';
+}
+
+
+function _addCtorParameters(classNode: ts.ClassDeclaration,
+                            angularImports: string[],
+                            refactor: TypeScriptFileRefactor) {
+  // For every classes with constructors, output the ctorParameters function which contains a list
+  // of injectable types.
+  const ctor = (
+    refactor.findFirstAstNode(classNode, ts.SyntaxKind.Constructor) as ts.ConstructorDeclaration);
+  if (!ctor) {
+    // A class can be missing a constructor, and that's _okay_.
+    return;
+  }
+
+  const params = Array.from(ctor.parameters).map(paramNode => {
+    switch (paramNode.type.kind) {
+      case ts.SyntaxKind.TypeReference:
+        return _ctorParameterFromTypeReference(paramNode, angularImports, refactor);
+      default:
+        return 'null';
+    }
+  });
+
+  const ctorParametersDecl = `static ctorParameters() { return [ ${params.join(', ')} ]; }`;
+  refactor.prependBefore(classNode.getLastToken(refactor.sourceFile), ctorParametersDecl);
+}
+
+
 function _removeDecorators(refactor: TypeScriptFileRefactor) {
-  // TODO: replace this by tsickle.
+  const angularImports: string[]
+    = refactor.findAstNodes(refactor.sourceFile, ts.SyntaxKind.ImportDeclaration)
+      .map((node: ts.ImportDeclaration) => _angularImportsFromNode(node, refactor.sourceFile))
+      .reduce((acc: string[], current: string[]) => acc.concat(current), []);
+
   // Find all decorators.
-  // refactor.findAstNodes(refactor.sourceFile, ts.SyntaxKind.Decorator)
-  //   .forEach(d => refactor.removeNode(d));
+  refactor.findAstNodes(refactor.sourceFile, ts.SyntaxKind.Decorator)
+    .forEach(node => {
+      // First, add decorators to classes to the classes array.
+      if (node.parent) {
+        const declarations = refactor.findAstNodes(node.parent,
+          ts.SyntaxKind.ClassDeclaration, false, 1);
+        if (declarations.length > 0) {
+          _addCtorParameters(declarations[0] as ts.ClassDeclaration, angularImports, refactor);
+        }
+      }
+
+      refactor.findAstNodes(node, ts.SyntaxKind.CallExpression)
+        .filter((node: ts.CallExpression) => {
+          const fnName = node.expression.getText(refactor.sourceFile);
+          if (fnName.indexOf('.') != -1) {
+            // Since this is `a.b`, see if it's the same namespace as a namespace import.
+            return angularImports.indexOf(fnName.replace(/\..*$/, '') + '.') != -1;
+          } else {
+            return angularImports.indexOf(fnName) != -1;
+          }
+        })
+        .forEach(() => refactor.removeNode(node));
+    });
 }
 
 
@@ -161,7 +289,7 @@ function _replaceResources(refactor: TypeScriptFileRefactor): void {
 
 
 function _checkDiagnostics(refactor: TypeScriptFileRefactor) {
-  const diagnostics = refactor.getDiagnostics();
+  const diagnostics: ts.Diagnostic[] = refactor.getDiagnostics();
 
   if (diagnostics.length > 0) {
     const message = diagnostics
@@ -176,10 +304,25 @@ function _checkDiagnostics(refactor: TypeScriptFileRefactor) {
 }
 
 
+/**
+ * Recursively calls diagnose on the plugins for all the reverse dependencies.
+ * @private
+ */
+function _diagnoseDeps(reasons: ModuleReason[], plugin: AotPlugin, checked: Set<string>) {
+  reasons
+    .filter(reason => reason && reason.module && reason.module instanceof NormalModule)
+    .filter(reason => !checked.has(reason.module.resource))
+    .forEach(reason => {
+      checked.add(reason.module.resource);
+      plugin.diagnose(reason.module.resource);
+      _diagnoseDeps(reason.module.reasons, plugin, checked);
+    });
+}
+
+
 // Super simple TS transpiler loader for testing / isolated usage. does not type check!
-export function ngcLoader(source: string) {
-  this.cacheable();
-  const cb: any = this.async();
+export function ngcLoader(this: LoaderContext & { _compilation: any }) {
+  const cb = this.async();
   const sourceFileName: string = this.resourcePath;
 
   const plugin = this._compilation._ngToolsWebpackPluginInstance as AotPlugin;
@@ -202,7 +345,13 @@ export function ngcLoader(source: string) {
       })
       .then(() => {
         if (plugin.typeCheck) {
-          _checkDiagnostics(refactor);
+          // Check all diagnostics from this and reverse dependencies also.
+          if (!plugin.firstRun) {
+            _diagnoseDeps(this._module.reasons, plugin, new Set<string>());
+          }
+          // We do this here because it will throw on error, resulting in rebuilding this file
+          // the next time around if it changes.
+          plugin.diagnose(sourceFileName);
         }
       })
       .then(() => {