+++ tests/e2e/tests/generate/module/module-routing-child-folder.ts
@@ -5,7 +5,7 @@ import { expectFileToExist } from '../../../utils/fs';
 import { expectToFail } from '../../../utils/utils';
 
 
-const Promise = require('angular-cli/ember-cli/lib/ext/promise');
+const Promise = require('@angular/cli/ember-cli/lib/ext/promise');
 
 export default function () {
   const root = process.cwd();