+++ tests/e2e_runner.ts
@@ -25,7 +25,7 @@ Error.stackTraceLimit = Infinity;
  * Here's a short description of those flags:
  *   --debug          If a test fails, block the thread so the temporary directory isn't deleted.
  *   --noproject      Skip creating a project or using one.
- *   --nolink         Skip linking your local angular-cli directory. Can save a few seconds.
+ *   --nolink         Skip linking your local @angular/cli directory. Can save a few seconds.
  *   --ng-sha=SHA     Use a specific ng-sha. Similar to nightly but point to a master SHA instead
  *                    of using the latest.
  *   --nightly        Install angular nightly builds over the test project.